# openapi.yaml
openapi: 3.0.3
info:
  title: Authentication Microservice API
  description: |
    A production-ready authentication microservice with comprehensive security features, 
    RBAC, JWT tokens, MFA, and GDPR compliance.
    
    ## Features
    - User registration and login with email verification
    - Role-based access control (Users, Providers, Admins)
    - JWT token management with access & refresh tokens
    - Multi-factor authentication using TOTP
    - Provider registration and verification workflow
    - GDPR compliance with data export and deletion
    - Comprehensive audit logging
    - Rate limiting and security headers
    
  version: 1.0.0
  contact:
    name: Auth System Support
    email: support@yourdomain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.yourdomain.com/api/v1
    description: Production server
  - url: https://staging-api.yourdomain.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users (Admin)
      summary: Get all users
      description: Get list of all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [user, provider, admin]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: isVerified
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Users list retrieved
  /users/{id}:
    get:
      tags:
        - Users (Admin)
      summary: Get user by ID
      description: Get specific user details (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User retrieved successfully

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...